FILE: creates and implements update services
