+++ libs/menu-matriarch/ingredients/data-access/src/lib/ingredient-type-data.service.ts
@@ -8,6 +8,7 @@ import {
   BatchService,
   Endpoint,
   IngredientTypeDto,
+  UserUpdateService,
   createIngredientTypeDto,
 } from '@atocha/menu-matriarch/shared/data-access';
 import { IngredientType } from '@atocha/menu-matriarch/shared/util';
@@ -24,7 +25,8 @@ export class IngredientTypeDataService {
 
   constructor(
     private _batchService: BatchService,
-    private _dataService: DataService
+    private _dataService: DataService,
+    private _userUpdateService: UserUpdateService
   ) {}
 
   getIngredientType(id: string): Observable<IngredientTypeDto | undefined> {
@@ -57,7 +59,7 @@ export class IngredientTypeDataService {
         data: createIngredientTypeDto({ id, uid, ...ingredientType }),
       })
       .update(
-        this._batchService.getUserUpdate({
+        this._userUpdateService.getUpdate({
           uid,
           ingredientTypeIdToAdd: id,
         })
@@ -82,7 +84,7 @@ export class IngredientTypeDataService {
     const batch = this._batchService.createBatch();
 
     batch.delete(this._endpoint, ingredientType.id).update(
-      this._batchService.getUserUpdate({
+      this._userUpdateService.getUpdate({
         uid: ingredientType.uid,
         ingredientTypeIdToDelete: ingredientType.id,
       })
