+++ libs/menu-matriarch/shared/data-access/src/index.ts
@@ -1,4 +1,3 @@
-export * from './lib/batch.service';
 export * from './lib/endpoint.enum';
 export * from './lib/error.service';
 export * from './lib/local-storage-key.enum';
@@ -18,3 +17,12 @@ export * from './lib/mappers/map-dish-dto-to-dish';
 export * from './lib/mappers/map-ingredient-type-dto-to-ingredient-type';
 export * from './lib/mappers/map-meal-dto-to-meal';
 export * from './lib/mappers/map-menu-dto-to-menu';
+
+export * from './lib/updaters/batch.service';
+export * from './lib/updaters/dish-update.service';
+export * from './lib/updaters/ingredient-type-update.service';
+export * from './lib/updaters/ingredient-update.service';
+export * from './lib/updaters/meal-update.service';
+export * from './lib/updaters/menu-update.service';
+export * from './lib/updaters/tag-update.service';
+export * from './lib/updaters/user-update.service';
